Revert "Event templates (#1230)"
This reverts commit 595229286a53d195f2141083f750dff2e6f2746d.
