+++ .eslintrc.js
@@ -18,12 +18,6 @@ module.exports = Object.assign({}, sharedConfigs, {
                 'no-empty-pattern': 0,
             },
         },
-        {
-            files: ['*.d.ts'],
-            rules: {
-                'spaced-comment': 0,
-            },
-        },
     ],
 });
