+++ client/actions/events/api.js
@@ -1431,50 +1431,6 @@ const receiveCalendars = (calendars) => ({
     payload: calendars,
 });
 
-const fetchEventTemplates = () => (dispatch, getState, {api}) => {
-    api('recent_events_template').query()
-        .then((res) => {
-            dispatch({type: EVENTS.ACTIONS.RECEIVE_EVENT_TEMPLATES, payload: res._items});
-        });
-};
-
-const createEventTemplate = (itemId) => (dispatch, getState, {api, modal}) => {
-    modal.prompt(gettext('Template name')).then((templateName) => {
-        api('events_template').query({
-            where: {
-                template_name: {
-                    $regex: templateName,
-                    $options: 'i',
-                },
-            },
-        })
-            .then((res) => {
-                const doSave = () => {
-                    api('events_template').save({
-                        template_name: templateName,
-                        based_on_event: itemId,
-                    })
-                        .then(() => {
-                            dispatch(fetchEventTemplates());
-                        });
-                };
-
-                const templateAlreadyExists = res._meta.total !== 0;
-
-                if (templateAlreadyExists) {
-                    modal.confirm(gettext(
-                        'Template already exists. Do you want to overwrite it?'
-                    ))
-                        .then(() => {
-                            api.remove(res._items[0], {}, 'events_template').then(doSave);
-                        });
-                } else {
-                    doSave();
-                }
-            });
-    });
-};
-
 // eslint-disable-next-line consistent-this
 const self = {
     loadEventsByRecurrenceId,
@@ -1512,8 +1468,6 @@ const self = {
     receiveCalendars,
     fetchEventFiles,
     removeFile,
-    fetchEventTemplates,
-    createEventTemplate,
 };
 
 export default self;
