+++ client/components/Events/EventItem.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {connect} from 'react-redux';
 import {get} from 'lodash';
 import {Label} from '../';
 import {EVENTS, MAIN, ICON_COLORS, WORKFLOW_STATE} from '../../constants';
@@ -17,10 +16,9 @@ import {
 } from '../../utils';
 import {gettext} from '../../utils/gettext';
 import {renderFields} from '../fields';
-import {getDeployConfig} from '../../selectors/config';
 
 
-export class EventItemComponent extends React.Component {
+export class EventItem extends React.Component {
     constructor(props) {
         super(props);
         this.state = {hover: false};
@@ -46,7 +44,7 @@ export class EventItemComponent extends React.Component {
             return null;
         }
 
-        const {session, privileges, item, lockedItems, calendars, deployConfig} = this.props;
+        const {session, privileges, item, lockedItems, calendars} = this.props;
         const callBacks = {
             [EVENTS.ITEM_ACTIONS.EDIT_EVENT.actionName]:
                 this.props[EVENTS.ITEM_ACTIONS.EDIT_EVENT.actionName].bind(null, item, true),
@@ -76,14 +74,10 @@ export class EventItemComponent extends React.Component {
                 this.props[EVENTS.ITEM_ACTIONS.UPDATE_REPETITIONS.actionName].bind(null, item),
             [EVENTS.ITEM_ACTIONS.ASSIGN_TO_CALENDAR.actionName]:
                 this.props[EVENTS.ITEM_ACTIONS.ASSIGN_TO_CALENDAR.actionName],
-            [EVENTS.ITEM_ACTIONS.SAVE_AS_TEMPLATE.actionName]:
-                this.props[EVENTS.ITEM_ACTIONS.SAVE_AS_TEMPLATE.actionName],
             [EVENTS.ITEM_ACTIONS.MARK_AS_COMPLETED.actionName]:
                 this.props[EVENTS.ITEM_ACTIONS.MARK_AS_COMPLETED.actionName].bind(null, item),
         };
-        const itemActions = eventUtils.getEventActions(
-            {item, session, privileges, lockedItems, callBacks, calendars, deployConfig}
-        );
+        const itemActions = eventUtils.getEventActions({item, session, privileges, lockedItems, callBacks, calendars});
 
         if (get(itemActions, 'length', 0) === 0) {
             return null;
@@ -223,7 +217,7 @@ export class EventItemComponent extends React.Component {
     }
 }
 
-EventItemComponent.propTypes = {
+EventItem.propTypes = {
     item: PropTypes.object.isRequired,
     onItemClick: PropTypes.func.isRequired,
     lockedItems: PropTypes.object.isRequired,
@@ -240,7 +234,6 @@ EventItemComponent.propTypes = {
     listFields: PropTypes.object,
     refNode: PropTypes.func,
     active: PropTypes.bool,
-    deployConfig: PropTypes.object.isRequired,
     [EVENTS.ITEM_ACTIONS.DUPLICATE.actionName]: PropTypes.func,
     [EVENTS.ITEM_ACTIONS.CREATE_PLANNING.actionName]: PropTypes.func,
     [EVENTS.ITEM_ACTIONS.CREATE_AND_OPEN_PLANNING.actionName]: PropTypes.func,
@@ -255,14 +248,6 @@ EventItemComponent.propTypes = {
     [EVENTS.ITEM_ACTIONS.ASSIGN_TO_CALENDAR.actionName]: PropTypes.func,
 };
 
-EventItemComponent.defaultProps = {
+EventItem.defaultProps = {
     togglePlanningItem: () => { /* no-op */ },
 };
-
-function mapStateToProps(state) {
-    return {
-        deployConfig: getDeployConfig(state),
-    };
-}
-
-export const EventItem = connect(mapStateToProps)(EventItemComponent);
