+++ client/components/Main/ItemEditor/EditorItemActions.jsx
@@ -1,30 +1,22 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {connect} from 'react-redux';
 
 import {ITEM_TYPE, EVENTS, PLANNING} from '../../../constants';
 import {getItemType, eventUtils, planningUtils} from '../../../utils';
-import eventsApi from '../../../actions/events/api';
-import * as allActions from '../../../actions';
 
 import {ItemActionsMenu} from '../../index';
-import {getDeployConfig} from '../../../selectors/config';
 
-const EditorItemActionsComponent = (props) => {
-    const {
-        item,
-        event,
-        session,
-        privileges,
-        lockedItems,
-        itemActions,
-        contentTypes,
-        itemManager,
-        autoSave,
-        dispatch,
-        deployConfig,
-    } = props;
-
+export const EditorItemActions = ({
+    item,
+    event,
+    session,
+    privileges,
+    lockedItems,
+    itemActions,
+    contentTypes,
+    itemManager,
+    autoSave,
+}) => {
     const itemType = getItemType(item);
     const withMultiPlanningDate = true;
     let actions = [], callBacks;
@@ -86,14 +78,6 @@ const EditorItemActionsComponent = (props) => {
                             itemActions[EVENTS.ITEM_ACTIONS.UPDATE_REPETITIONS.actionName](item)
                         ))
                 ),
-            [EVENTS.ITEM_ACTIONS.SAVE_AS_TEMPLATE.actionName]:
-                () => {
-                    const message = gettext('Save changes before creating a template?');
-
-                    dispatch(allActions.main.openActionModalFromEditor(item, message, (updatedItem) => {
-                        dispatch(eventsApi.createEventTemplate(updatedItem._id));
-                    }));
-                },
             [EVENTS.ITEM_ACTIONS.MARK_AS_COMPLETED.actionName]:
                 () => (
                     autoSave.flushAutosave()
@@ -109,7 +93,6 @@ const EditorItemActionsComponent = (props) => {
             lockedItems,
             callBacks,
             withMultiPlanningDate,
-            deployConfig,
         });
     }
 
@@ -167,7 +150,7 @@ const EditorItemActionsComponent = (props) => {
     />);
 };
 
-EditorItemActionsComponent.propTypes = {
+EditorItemActions.propTypes = {
     item: PropTypes.object,
     event: PropTypes.object,
     session: PropTypes.object,
@@ -177,14 +160,4 @@ EditorItemActionsComponent.propTypes = {
     contentTypes: PropTypes.array,
     itemManager: PropTypes.object,
     autoSave: PropTypes.object,
-    dispatch: PropTypes.func,
-    deployConfig: PropTypes.object,
 };
-
-function mapStateToProps(state) {
-    return {
-        deployConfig: getDeployConfig(state),
-    };
-}
-
-export const EditorItemActions = connect(mapStateToProps)(EditorItemActionsComponent);
