+++ client/components/fields/calendars.jsx
@@ -22,10 +22,7 @@ export const calendars = ({item, calendars, grow}) => {
 };
 
 calendars.propTypes = {
-    item: PropTypes.shape({
-        description_text: PropTypes.string,
-        calendars: PropTypes.array,
-    }).isRequired,
+    item: PropTypes.object.isRequired,
     calendars: PropTypes.array,
     grow: PropTypes.bool,
 };
