+++ client/constants/modals.js
@@ -12,6 +12,5 @@ export const MODALS = {
     UNLOCK_FEATURED_STORIES: 'UNLOCK_FEATURED_STORIES',
     SELECT_DESK_TEMPLATE: 'SELECT_DESK_TEMPLATE',
     MANAGE_EVENTS_PLANNING_FILTERS: 'MANAGE_EVENTS_PLANNING_FILTERS',
-    MANAGE_EVENT_TEMPLATES: 'MANAGE_EVENT_TEMPLATES',
     EXPORT_AS_ARTICLE: 'EXPORT_AS_ARTICLE',
 };
