+++ client/services/PlanningStoreService.js
@@ -29,7 +29,6 @@ export class PlanningStoreService {
         $q,
         $interpolate,
         search,
-        modal,
         preferencesService
     ) {
         this.$rootScope = $rootScope;
@@ -55,7 +54,6 @@ export class PlanningStoreService {
         this.$q = $q;
         this.$interpolate = $interpolate;
         this.search = search;
-        this.modal = modal;
         this.preferencesService = preferencesService;
 
         this.onSessionChanged = this.onSessionChanged.bind(this);
@@ -163,7 +161,6 @@ export class PlanningStoreService {
                         authoringWorkspace: this.authoringWorkspace,
                         $interpolate: this.$interpolate,
                         search: this.search,
-                        modal: this.modal,
                         config: this.config,
                         preferencesService: this.preferencesService,
                         $rootScope: this.$rootScope,
@@ -331,6 +328,5 @@ PlanningStoreService.$inject = [
     '$q',
     '$interpolate',
     'search',
-    'modal',
     'preferencesService',
 ];
