--- client/typings/refs.d.ts
@@ -1,1 +1,0 @@
-/// <reference path='../../node_modules/superdesk-core/scripts/core/superdesk-api.d.ts' />
