+++ client/utils/events.js
@@ -580,16 +580,7 @@ const getMultiDayPlanningActions = (item, actions, createPlanning, createAndOpen
     }
 };
 
-const getEventActions = ({
-    item,
-    session,
-    privileges,
-    lockedItems,
-    callBacks,
-    withMultiPlanningDate,
-    calendars,
-    deployConfig,
-}) => {
+const getEventActions = ({item, session, privileges, lockedItems, callBacks, withMultiPlanningDate, calendars}) => {
     if (!isExistingItem(item)) {
         return [];
     }
@@ -611,10 +602,6 @@ const getEventActions = ({
         EVENTS.ITEM_ACTIONS.CONVERT_TO_RECURRING.actionName,
         EVENTS.ITEM_ACTIONS.MARK_AS_COMPLETED.actionName,
     ];
-
-    if (deployConfig.event_templates_enabled === true) {
-        alllowedCallBacks.push(EVENTS.ITEM_ACTIONS.SAVE_AS_TEMPLATE.actionName);
-    }
 
     if (isExpired && !privileges[PRIVILEGES.EDIT_EXPIRED]) {
         alllowedCallBacks = [EVENTS.ITEM_ACTIONS.DUPLICATE.actionName];
