+++ server/planning/common.py
@@ -102,12 +102,6 @@ def planning_auto_assign_to_workflow(current_app=None):
     return app.config.get('PLANNING_AUTO_ASSIGN_TO_WORKFLOW', False)
 
 
-def event_templates_enabled(current_app=None):
-    if current_app is not None:
-        return current_app.config.get('PLANNING_EVENT_TEMPLATES_ENABLED', False)
-    return app.config.get('PLANNING_EVENT_TEMPLATES_ENABLED', False)
-
-
 def get_long_event_duration_threshold(current_app=None):
     if current_app is not None:
         return current_app.config.get('LONG_EVENT_DURATION_THRESHOLD', -1)
